openweather/requirements.txt
oauth2client==3.0.0
GitPython==3.1.24
folium==0.2.1
urllib3
selenium
random-user-agent
webdriver_manager
webdriver-setup
pyvirtualdisplay
pandas
sqlalchemy
apache-airflow[amazon]
apache-airflow-providers-amazon
apache-airflow-providers-microsoft-azure
azure-keyvault-secrets
azure-identity
# azure
# apache-airflow-backport-providers-microsoft-azure
pyarrow
fastparquet
numpy
crypto==1.4.1
pycryptodome
sai-airflow-plugins
